Match Sfx_PlayFromObjectEx 96.1->100% (audio): #21 inverted ||-form for the found=NULL arm (NULL arm laid first), (int) on the matchCount==3 compare, and recipe #105 K&R definition for Sfx_AllocObjectChannel (s16 a under K&R + int-typed prototype: caller passes raw lhz with no extsh while the callee body keeps the s16 extension semantics; the dead f32 pitch param becomes double per K&R promotion, prototype updated to match -- caller codegen identical since lfs f1 passes through). Alloc holds its 90.35 baseline exactly; zero regressions project-wide

+++ include/main/engine_shared.h
@@ -1496,7 +1496,7 @@ void *_audioAlloc(u32 size);
 void Music_ChannelLoadedCallback(MusicBank *bank, MusicChannel *channel, MusicTrigParam *trigger);
 int Sfx_ReadTriggerParams(SfxTriggerFull *trigger, u16 *outSfxId, u8 *outVol, f32 *outF6, f32 *outF7, f32 *outF8, int *outI9, int *outI10, int *outI11);
 SfxTrigger *Sfx_FindTrigger(u16 id);
-SfxObjectChannel *Sfx_AllocObjectChannel(s16 a, int b, f32 pitch, int c, int d);
+SfxObjectChannel *Sfx_AllocObjectChannel(int a, int b, double pitch, int c, int d);
 void Sfx_RotateVectorByAngles(s16 angX, s16 angY, s16 angZ, f32 *v);
 f32 Sfx_GetListenerRelativeDistance(f32 *soundPos, f32 *outDelta);
 void AudioStream_StopAll(void);
